Splitting out test fixtures that were is same file.
